src/apps/origen/partner/resolvers/origen-get-partners.resolver.spec.ts
import { Test, TestingModule } from '@nestjs/testing';

// custom items
import { OrigenGetPartnersResolver } from './origen-get-partners.resolver'; 
import { ICommandBus } from '@hades/shared/domain/bus/command-bus';
import { IQueryBus } from '@hades/shared/domain/bus/query-bus';
import { partners } from '@hades/origen/partner/infrastructure/seeds/partner.seed';

describe('OrigenGetPartnersResolver', () => 
{
    let resolver:   OrigenGetPartnersResolver;
    let queryBus: IQueryBus;
    let commandBus: ICommandBus;

    beforeAll(async () => 
    {
        const module: TestingModule = await Test.createTestingModule({
            providers: [
                OrigenGetPartnersResolver,
                {
                    provide: IQueryBus,
                    useValue: {
                        ask: () => {},
                    }
                },
                {
                    provide: ICommandBus,
                    useValue: {
                        dispatch: () => {},
                    }
                },
            ]
        }).compile();

        resolver    = module.get<OrigenGetPartnersResolver>(OrigenGetPartnersResolver);
        queryBus    = module.get<IQueryBus>(IQueryBus);
        commandBus  = module.get<ICommandBus>(ICommandBus);
    });

    test('OrigenGetPartnersResolver should be defined', () => 
    {
        expect(resolver).   toBeDefined();
    });

    describe('main', () => 
    {
        test('OrigenGetPartnersResolver should be defined', () => 
        {
            expect(resolver).   toBeDefined();
        });

        test('should return a partners', async () => 
        {
            jest.spyOn(queryBus, 'ask').mockImplementation(() => new Promise(resolve => resolve(partners)));
            expect(await resolver.main()).toBe(partners);
        });
    });
});